requirements.txt
amqp==2.6.1
asgiref==3.5.2
asttokens==2.2.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cachetools==5.2.0
celery==4.4.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.3
django-allauth==0.51.0
django-braces==1.15.0
django-cors-headers==3.13.0
django-filter==22.1
django-oauth-toolkit==2.2.0
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
executing==1.2.0
google-auth==2.15.0
icecream==2.1.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
kombu==4.6.11
MarkupSafe==2.1.1
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==21.3
prompt-toolkit==3.0.33
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
vine==1.3.0
wcwidth==0.2.5
wrapt==1.14.1
